Better readability on prints
with f"{}" its easier to see what the print means, i guess?
i know its such a small thing, but hey, this library really helps me on a day to day basis, so i want to be a part of it! even if its just a little!
(if f"{}" doesn't work on sys.exit, im sorry!)
take it easy on me, my first pull request / change to a project.